Some tidying done, no tests yet..
