+++ app/uk/gov/hmrc/softdrinksindustrylevystub/controllers/SubscriptionController.scala
@@ -21,18 +21,13 @@ import javax.inject.{Inject, Singleton}
 import play.api.libs.json.{JsValue, Json}
 import play.api.mvc._
 import uk.gov.hmrc.play.microservice.controller.BaseController
-import uk.gov.hmrc.softdrinksindustrylevystub.models.DesSubmissionRequest
-import uk.gov.hmrc.softdrinksindustrylevystub.models.etmp.createsub.{CreateSubscriptionRequest, CreateSubscriptionResponse}
+import uk.gov.hmrc.softdrinksindustrylevystub.models.etmp.createsub.CreateSubscriptionRequest
 import uk.gov.hmrc.softdrinksindustrylevystub.services.DesSubmissionService
 
 import scala.concurrent.Future
 
 @Singleton
 class SubscriptionController @Inject()(desSubmissionService: DesSubmissionService) extends BaseController {
-
-  def submitHello(): Action[JsValue] = Action.async(parse.json) { implicit request =>
-    withJsonBody[DesSubmissionRequest](_ => Future.successful(Ok(Json.toJson(desSubmissionService.buildResponse()))))
-  }
 
   def createSubscription(): Action[JsValue] = Action.async(parse.json) { implicit request =>
     withJsonBody[CreateSubscriptionRequest](data =>
